sqlalchemy/prac2/requirements.txt
sqlalchemy
asyncpg
python-dotenv
questionary 